P14-26 Created website framework
Laid out the general framework of the website.  This includes most of the pages we will require, and some basic styling for the website header.  Note: most styles are temporary.

Most pages lack content, however all pages have descriptions for what they need to contain, and unimplemented functionality.

This also includes dummy data, as this website is not hooked up to any database yet.