Added Prisma Client
I have laid out the groundwork for the database.  This includes creating the initial schema which stores authenticated users.
I have also updated the header to show whether a user is logged in or out.
I have removed username/password support for instructors, as Next-Auth discourages this for security reasons.